Avoid importing by package name (#67)
We want to integrate this code into the base python API, forming a unified package. This inherently changes the package name: here, it is cvc5_z3py_compat, while it is cvc5.pythonic in the actual package. For this to work, we can only use from . import ..., but not from cvc5_z3py_compat import ... in any code that is actually relocated.